requirements.txt
Django==3.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
dj-rest-auth==2.2.1
django-cors-headers==3.10.1
